requirements.txt
aniso8601
annotated-types
anyio
blinker
certifi
charset-normalizer
click
colorama
contourpy
cycler
filelock
Flask
Flask-Cors
Flask-RESTful
Flask-SQLAlchemy
fonttools
greenlet
h11
httpcore
httpx
huggingface-hub
humanize
idna
itsdangerous
Jinja2
joblib
keybert
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
mdurl
mpmath
networkx
nltk
numpy
ollama
packaging
pillow
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
razorpay
regex
requests
rich
safetensors
scikit-learn
scipy
sentence-transformers
six
sniffio
SQLAlchemy
sympy
threadpoolctl
tokenizers
torch
tqdm
transformers
typing_extensions
urllib3
Werkzeug
google-generativeai
tensorflow-intel
langchain
python-dotenv
langchain-google-genai